AlignmentUI-main/requirements.txt
absl-py==2.3.0
albucore==0.0.24
albumentations==2.0.8
altgraph==0.17.4
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
clip @ git+https://github.com/ultralytics/CLIP.git@a6238961f9fab312d99c7709f8c7485b3f9a9b50
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
ftfy==6.3.1
future==1.0.0
GPUtil==1.4.0
greenlet==3.2.2
grpcio==1.72.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
lap==0.5.12
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
pyparsing==3.2.3
pypylon==4.2.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.17.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzbar==0.1.9
pyzmq==26.4.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
regex==2024.11.6
reportlab==4.4.3
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
serial==0.0.97
simsimd==6.4.9
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.167
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
zxing-cpp==2.3.0
